De-duplicate Dutch tests
Test data that applies to both nl-BE and nl-NL used to be stored in a
common file (6f1b5ca) that was then "included" in the test files of both
tables. Later when json tests were migrated to yaml (139d8bc5) the file
was split out, thereby duplicating test data. Now with the ability to
test multiple tables with the same set of tests in the YAML framework,
we can de-duplicate that data again.